supervisor dashboard seeder updates
