+++ backend/src/database/seed/create-personnel.ts
@@ -1,62 +1,35 @@
 import { faker } from '@faker-js/faker';
-import { divisionsAndMinistries } from './const';
-import { CreatePersonnelDTO } from '../personnel';
-import { DriverLicense } from './enums/driver-license.enum';
-import { Ministry } from './enums/ministry.enum';
-import { Status } from './enums/status.enum';
-import { UnionMembership } from './enums/union-membership.enum';
+import { divisionsAndMinistries } from '../../common/const';
+import { CreatePersonnelDTO } from '../../personnel';
+import { DriverLicense } from '../../common/enums/driver-license.enum';
+import { Ministry } from '../../common/enums/ministry.enum';
+import { Status } from '../../common/enums/status.enum';
+import { UnionMembership } from '../../common/enums/union-membership.enum';
 import {
   createTools,
   createCertifications,
   createLanguages,
   availability,
-} from './seed-common';
-import { CreatePersonnelBcwsDTO } from '../bcws/dto/create-bcws-personnel.dto';
-import { BcwsRoleEntity } from '../database/entities/bcws/bcws-role.entity';
-import {
-  EmcrFunctionEntity,
-  EmcrTrainingEntity,
-} from '../database/entities/emcr';
-import { LocationEntity } from '../database/entities/location.entity';
-import { AvailabilityEntity } from '../database/entities/personnel/availability.entity';
-import { CertificationEntity } from '../database/entities/personnel/certifications.entity';
-import { ToolsEntity } from '../database/entities/personnel/tools.entity';
-import { createBCWShandler } from '../database/seed-bcws-partial';
-import { createEMCRhandler } from '../database/seed-emcr-partial';
-import { CreatePersonnelEmcrDTO } from '../emcr/dto';
+} from './helpers';
+import { LocationEntity } from '../entities/location.entity';
+import { AvailabilityEntity } from '../entities/personnel/availability.entity';
+import { CertificationEntity } from '../entities/personnel/certifications.entity';
+import { ToolsEntity } from '../entities/personnel/tools.entity';
 
-export const handler = (
+
+export const createPersonnelHandler = (
+  status: Status, 
   locations?: LocationEntity[],
   tools?: ToolsEntity[],
   certs?: CertificationEntity[],
-  roles?: BcwsRoleEntity[],
-  functions?: EmcrFunctionEntity[],
-  trainings?: EmcrTrainingEntity[],
 ): {
   personnelData: CreatePersonnelDTO;
-  bcwsData: CreatePersonnelBcwsDTO;
-  emcrData: CreatePersonnelEmcrDTO;
 } => {
-  const status =
-    Status[
-      faker.helpers.arrayElement([
-        Status.ACTIVE,
-        Status.INACTIVE,
-        Status.PENDING,
-      ])
-    ];
-
-  const dateApplied = faker.date.past();
+  
   const divisionAndMinistry = faker.helpers.arrayElement(
     divisionsAndMinistries,
   );
-  const { emcrData } = createEMCRhandler(
-    functions,
-    trainings,
-    status,
-    dateApplied,
-  );
-  const { bcwsData } = createBCWShandler(roles, status, dateApplied);
+  
 
   const personnelData: CreatePersonnelDTO = {
     homeLocation: faker.helpers.arrayElement(locations),
@@ -95,5 +68,5 @@ export const handler = (
     availability:
       status !== Status.PENDING ? (availability() as AvailabilityEntity[]) : [],
   };
-  return { personnelData, bcwsData, emcrData };
+  return { personnelData };
 };
