+++ backend/src/database/seed/seed-bcws.ts
@@ -1,7 +1,10 @@
-import { datasource } from './datasource';
-import { BcwsPersonnelEntity } from './entities/bcws';
-import { handler as dataHandler } from '../common/bcws-seed';
-import { PersonnelEntity } from './entities/personnel/personnel.entity';
+import { Status } from '../../common/enums';
+import { datasource } from '../datasource';
+import { BcwsPersonnelEntity } from '../entities/bcws';
+import { PersonnelEntity } from '../entities/personnel/personnel.entity';
+import { createBCWShandler } from './create-bcws';
+import { createPersonnelHandler } from './create-personnel';
+import { faker } from '@faker-js/faker';
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const handler = async () => {
   if (!datasource.isInitialized) {
@@ -14,14 +17,26 @@ export const handler = async () => {
   const roles = await datasource.query('SELECT * FROM bcws_role');
   const personnelRepo = datasource.getRepository(PersonnelEntity);
   const bcwsPersonnelRepo = datasource.getRepository(BcwsPersonnelEntity);
+  const dateApplied = faker.date.past();
 
+  const status =
+  Status[
+    faker.helpers.arrayElement([
+      Status.ACTIVE,
+      Status.INACTIVE,
+      Status.PENDING,
+    ])
+  ];
+
+  
   try {
-    const { personnelData, bcwsData } = dataHandler(
+    const { personnelData } = createPersonnelHandler(
+      status,
       locations,
-      roles,
       tools,
       certs,
     );
+    const { bcwsData } = createBCWShandler(roles, status, dateApplied);
 
     const person = await personnelRepo.save(
       personnelRepo.create(
@@ -41,12 +56,17 @@ export const handler = async () => {
     );
 
     for (let i = 0; i < 50; i++) {
-      const { personnelData, bcwsData } = dataHandler(
+      const dateApplied = faker.date.past();
+      const { personnelData } = createPersonnelHandler(
+        status,
         locations,
-        roles,
         tools,
         certs,
       );
+      
+      
+      
+      const { bcwsData } = createBCWShandler(roles, status, dateApplied);
 
       const person = await personnelRepo.save(
         personnelRepo.create(new PersonnelEntity(personnelData)),
