+++ backend/src/database/seed/seed-emcr.ts
@@ -1,8 +1,11 @@
-import { datasource } from './datasource';
-import { handler as dataHandler } from '../common/emcr-seed';
-import { EmcrPersonnelEntity } from './entities/emcr';
-import { LocationEntity } from './entities/location.entity';
-import { PersonnelEntity } from './entities/personnel/personnel.entity';
+import { datasource } from '../datasource';
+import { EmcrPersonnelEntity } from '../entities/emcr';
+import { LocationEntity } from '../entities/location.entity';
+import { PersonnelEntity } from '../entities/personnel/personnel.entity';
+import { createEMCRhandler } from './create-emcr';
+import { createPersonnelHandler } from './create-personnel';
+import { faker } from '@faker-js/faker';
+import { Status } from '../../common/enums';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const handler = async () => {
@@ -19,16 +22,28 @@ export const handler = async () => {
   const emcrPersonnelRepo = datasource.getRepository(EmcrPersonnelEntity);
 
   try {
-    const { personnelData, emcrData } = dataHandler(
-      locations.map((itm) => ({
-        locationName: itm.locationName,
-        region: itm.region,
-        id: itm.id,
-      })),
-      functions,
-      trainings,
+    const status =
+      Status[
+        faker.helpers.arrayElement([
+          Status.ACTIVE,
+          Status.INACTIVE,
+          Status.PENDING,
+        ])
+      ];
+
+    const dateApplied = faker.date.past();
+    const { personnelData } = createPersonnelHandler(
+      status,
+      locations,
       tools,
       certs,
+    );
+
+    const { emcrData } = createEMCRhandler(
+      functions,
+      trainings,
+      status,
+      dateApplied,
     );
 
     const person = await personnelRepo.save(
@@ -46,47 +61,30 @@ export const handler = async () => {
         new EmcrPersonnelEntity({ ...emcrData, personnelId: person.id }),
       ),
     );
-    const personTwo = await personnelRepo.save(
-      personnelRepo.create(
-        new PersonnelEntity({
-          ...personnelData,
-          email: 'emcr-coordinator@gov.bc.ca',
-          supervisorEmail: 'supervisor@gov.bc.ca',
-        }),
-      ),
-    );
 
-    await emcrPersonnelRepo.save(
-      emcrPersonnelRepo.create(
-        new EmcrPersonnelEntity({ ...emcrData, personnelId: personTwo.id }),
-      ),
-    );
-
-    const personThree = await personnelRepo.save(
-      personnelRepo.create(
-        new PersonnelEntity({
-          ...personnelData,
-          email: 'EMCR.LogisticsOfficerTest@gov.bc.ca',
-          supervisorEmail: 'BCWS.CoordinatorTest@gov.bc.ca',
-        }),
-      ),
-    );
-    await emcrPersonnelRepo.save(
-      emcrPersonnelRepo.create(
-        new EmcrPersonnelEntity({ ...emcrData, personnelId: personThree.id }),
-      ),
-    );
     for (let i = 0; i < 50; i++) {
-      const { personnelData, emcrData } = dataHandler(
-        locations.map((itm) => ({
-          locationName: itm.locationName,
-          region: itm.region,
-          id: itm.id,
-        })),
-        functions,
-        trainings,
+      const status =
+        Status[
+          faker.helpers.arrayElement([
+            Status.ACTIVE,
+            Status.INACTIVE,
+            Status.PENDING,
+          ])
+        ];
+
+      const dateApplied = faker.date.past();
+      const { personnelData } = createPersonnelHandler(
+        status,
+        locations,
         tools,
         certs,
+      );
+
+      const { emcrData } = createEMCRhandler(
+        functions,
+        trainings,
+        status,
+        dateApplied,
       );
 
       const person = await personnelRepo.save(
